Move configuration area to XDG_RUNTIME_DIR
The Eclipse configuration area [1] needs to be cleared whenever the Eclipse
framework is updated; otherwise Portfolio Performance crashes on startup [2].

Set the configuration path to
${XDG_RUNTIME_DIR}/app/info.portfolio_performance.PortfolioPerformance/configuration,
which is guaranteed to be cleared whenever the user logs out.

This location is on a tmpfs, which could be problematic if large files are
created there in the future. However, for the moment, we don't have a better
option; see bug [3]

Fixes #66

[1]
https://help.eclipse.org/latest/topic/org.eclipse.platform.doc.isv/reference/misc/runtime-options.html#osgiconfigurationarea
[2] https://github.com/portfolio-performance/portfolio/issues/3418
[3] https://github.com/flatpak/flatpak/issues/3438